Fix broken RPM source files preparation
The source files that were copied to MOCK_CHROOT_BUILD_DIR got overwritten when
running the command to build the source RPM.

This patch copies those files to the directory where the package sources
archives are generated, ensuring they are available when it is time to build
the RPMs.

To ensure consistent behavior, both the new and old YAML syntax have their
files copied to the archive directory.